# kingsAREGreat22/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Order Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  name: string;
  price: number;
  stock: number;
  sku: string;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  price: number;
  stock?: number;
  sku?: string;
}

export interface OrderItem {
  productId: string;
  name: string;
  price: number;
  quantity: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  NEW: "NEW",
  PACKED: "PACKED",
  SHIPPED: "SHIPPED",
  DELIVERED: "DELIVERED",
} as const;

export interface Order {
  id: string;
  customer: string;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  createdAt: string;
}

export type CreateOrderBodyItemsItem = {
  productId: string;
  quantity: number;
};

export interface CreateOrderBody {
  customer?: string;
  items: CreateOrderBodyItemsItem[];
}

export type DashboardStatsOrdersByStatus = { [key: string]: number };

export interface DashboardStats {
  totalProducts: number;
  totalOrders: number;
  openOrders: number;
  deliveredRevenue: number;
  stockValue: number;
  ordersByStatus: DashboardStatsOrdersByStatus;
}

export interface SuccessResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}