# kingsAREGreat22/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Order Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all products
 */
export const ListProductsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  price: zod.number(),
  stock: zod.number(),
  sku: zod.string(),
  createdAt: zod.string(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Add a product to inventory
 */
export const CreateProductBody = zod.object({
  name: zod.string(),
  price: zod.number(),
  stock: zod.number().optional(),
  sku: zod.string().optional(),
});

/**
 * @summary Remove a product
 */
export const DeleteProductParams = zod.object({
  id: zod.coerce.string(),
});

export const DeleteProductResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary List all orders
 */
export const ListOrdersResponseItem = zod.object({
  id: zod.string(),
  customer: zod.string(),
  items: zod.array(
    zod.object({
      productId: zod.string(),
      name: zod.string(),
      price: zod.number(),
      quantity: zod.number(),
    }),
  ),
  total: zod.number(),
  status: zod.enum(["NEW", "PACKED", "SHIPPED", "DELIVERED"]),
  createdAt: zod.string(),
});
export const ListOrdersResponse = zod.array(ListOrdersResponseItem);

/**
 * @summary Create a new order
 */
export const CreateOrderBody = zod.object({
  customer: zod.string().optional(),
  items: zod.array(
    zod.object({
      productId: zod.string(),
      quantity: zod.number(),
    }),
  ),
});

/**
 * @summary Advance order to the next status
 */
export const AdvanceOrderParams = zod.object({
  id: zod.coerce.string(),
});

export const AdvanceOrderResponse = zod.object({
  id: zod.string(),
  customer: zod.string(),
  items: zod.array(
    zod.object({
      productId: zod.string(),
      name: zod.string(),
      price: zod.number(),
      quantity: zod.number(),
    }),
  ),
  total: zod.number(),
  status: zod.enum(["NEW", "PACKED", "SHIPPED", "DELIVERED"]),
  createdAt: zod.string(),
});

/**
 * @summary Get dashboard summary statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalProducts: zod.number(),
  totalOrders: zod.number(),
  openOrders: zod.number(),
  deliveredRevenue: zod.number(),
  stockValue: zod.number(),
  ordersByStatus: zod.record(zod.string(), zod.number()),
});